v2.0.0 - Redesign, Complex Numbers, Custom Tensors (#38)
* Major version with breaking changes.

**Breaking Changes**:
- `onpzstream` has been split into `npzfilewriter` and `npzstringwriter`
- `inpzstream` has been split into `npzfilereader` and `npzstringreader`
- The library interface has been merged into a single `npy.h` header

Improvements:
- Support for complex numbers has been added
- The library design has been greatly simplified
- The example projects have been broken out into self-contained projects (which fetch
  the main project using `FetchContent`) and a `custom_tensors` example has been added
  showing a non-trivial custom tensor implementation
- Custom tensor support is greatly improved, and is now fully featured

Bugfixes:
- Fixed an issue with broken CRC checks on Windows

+++ test/generate_large_test.py
@@ -3,13 +3,15 @@ import os
 import numpy as np
 
 if __name__ == "__main__":
+    path = os.path.join(os.path.dirname(__file__), "..",
+                        "assets", "test", "test_large.npz")
+    if os.path.exists(path):
+        exit()
+
     test_int = np.arange(1000000).astype(np.int32)
     test_int = test_int.reshape(200, 5, 1000)
     test_float = np.arange(1000000).astype(np.float32)
     test_float = test_float.reshape(1000, 5, 20, 10)
-
-    path = os.path.join(os.path.dirname(__file__), "..",
-                        "assets", "test", "test_large.npz")
     np.savez(path, test_int=test_int, test_float=test_float)
 
     path = os.path.join(os.path.dirname(__file__), "..",
